@@ -1,6 +1,6 @@
 module github.com/caddyserver/caddy/v2
 
-go 1.17
+go 1.18
 
 require (
 	github.com/BurntSushi/toml v1.2.0
